+++ x/leverage/keeper/grpc_query.go
@@ -134,9 +134,12 @@ func (q Querier) MarketSummary(
 		AvailableCollateralize: availableCollateralize,
 	}
 
-	// Oracle price in response will be nil if it is unavailable
-	if oraclePrice, _, oracleErr := q.Keeper.TokenDefaultDenomPrice(ctx, req.Denom, false); oracleErr == nil {
+	// Oracle prices in response will be nil if it is unavailable
+	if oraclePrice, _, oracleErr := q.Keeper.TokenPrice(ctx, req.Denom, types.PriceModeSpot); oracleErr == nil {
 		resp.OraclePrice = &oraclePrice
+	}
+	if historicPrice, _, historicErr := q.Keeper.TokenPrice(ctx, req.Denom, types.PriceModeHistoric); historicErr == nil {
+		resp.OracleHistoricPrice = &historicPrice
 	}
 
 	return &resp, nil
@@ -199,22 +202,30 @@ func (q Querier) AccountSummary(
 	collateral := q.Keeper.GetBorrowerCollateral(ctx, addr)
 	borrowed := q.Keeper.GetBorrowerBorrows(ctx, addr)
 
-	suppliedValue, err := q.Keeper.TotalTokenValue(ctx, supplied, false)
+	// supplied value always uses spot prices
+	suppliedValue, err := q.Keeper.TotalTokenValue(ctx, supplied, types.PriceModeSpot)
 	if err != nil {
 		return nil, err
 	}
-	borrowedValue, err := q.Keeper.TotalTokenValue(ctx, borrowed, false)
+	// borrowed value here is shown using spot prices, but leverage logic instead uses
+	// the higher of spot or historic prices for each borrowed token when comparing it
+	// to borrow limit.
+	borrowedValue, err := q.Keeper.TotalTokenValue(ctx, borrowed, types.PriceModeSpot)
 	if err != nil {
 		return nil, err
 	}
+	// collateral value always uses spot prices
 	collateralValue, err := q.Keeper.CalculateCollateralValue(ctx, collateral)
 	if err != nil {
 		return nil, err
 	}
-	borrowLimit, err := q.Keeper.CalculateBorrowLimit(ctx, collateral, false)
+	// borrow limit shown here as it is used in leverage logic:
+	// using the lower of spot or historic prices for each collateral token
+	borrowLimit, err := q.Keeper.CalculateBorrowLimit(ctx, collateral)
 	if err != nil {
 		return nil, err
 	}
+	// liquidation always uses spot prices
 	liquidationThreshold, err := q.Keeper.CalculateLiquidationThreshold(ctx, collateral)
 	if err != nil {
 		return nil, err
@@ -277,12 +288,10 @@ func (q Querier) MaxWithdraw(
 	if req == nil {
 		return nil, status.Error(codes.InvalidArgument, "empty request")
 	}
-	if req.Address == "" {
-		return nil, status.Error(codes.InvalidArgument, "empty address")
+	if err := req.ValidateBasic(); err != nil {
+		return nil, err
 	}
-
 	ctx := sdk.UnwrapSDKContext(goCtx)
-
 	addr, err := sdk.AccAddressFromBech32(req.Address)
 	if err != nil {
 		return nil, err
@@ -298,24 +307,17 @@ func (q Querier) MaxWithdraw(
 	} else {
 		// Denom not specified
 		for _, t := range q.Keeper.GetAllRegisteredTokens(ctx) {
-			denoms = append(denoms, t.BaseDenom)
+			if !t.Blacklist {
+				denoms = append(denoms, t.BaseDenom)
+			}
 		}
 	}
 
 	for _, denom := range denoms {
-		maxCurrentWithdraw, err := q.Keeper.maxWithdraw(ctx, addr, denom, false)
+		uToken, err := q.Keeper.maxWithdraw(ctx, addr, denom)
 		if err != nil {
 			return nil, err
 		}
-		maxHistoricWithdraw, err := q.Keeper.maxWithdraw(ctx, addr, denom, true)
-		if err != nil {
-			return nil, err
-		}
-
-		uToken := sdk.NewCoin(
-			maxCurrentWithdraw.Denom,
-			sdk.MinInt(maxCurrentWithdraw.Amount, maxHistoricWithdraw.Amount),
-		)
 		if uToken.IsPositive() {
 			token, err := q.Keeper.ExchangeUToken(ctx, uToken)
 			if err != nil {
@@ -359,23 +361,17 @@ func (q Querier) MaxBorrow(
 	} else {
 		// Denom not specified
 		for _, t := range q.Keeper.GetAllRegisteredTokens(ctx) {
-			denoms = append(denoms, t.BaseDenom)
+			if !t.Blacklist {
+				denoms = append(denoms, t.BaseDenom)
+			}
 		}
 	}
 
 	for _, denom := range denoms {
-		currentMaxBorrow, err := q.Keeper.maxBorrow(ctx, addr, denom, false)
+		maxBorrow, err := q.Keeper.maxBorrow(ctx, addr, denom)
 		if err != nil {
 			return nil, err
 		}
-		historicMaxBorrow, err := q.Keeper.maxBorrow(ctx, addr, denom, true)
-		if err != nil {
-			return nil, err
-		}
-		maxBorrow := sdk.NewCoin(
-			currentMaxBorrow.Denom,
-			sdk.MinInt(currentMaxBorrow.Amount, historicMaxBorrow.Amount),
-		)
 		if maxBorrow.IsPositive() {
 			maxTokens = maxTokens.Add(maxBorrow)
 		}